+++ src/lazy_log/constants.py
@@ -1,0 +1,4 @@
+"""Module for constants used in the package."""
+
+PROG_NAME = "lazy-log-formatter"
+VENV_DIRS = {".venv", ".env"}